pre CV32F to CV64F migation

+++ include/headers.h
@@ -51,7 +51,7 @@ const std::string reset("\033[0m");
 #include <CGAL/Simple_cartesian.h>
 #include <CGAL/linear_least_squares_fitting_3.h>
 
-typedef CGAL::Simple_cartesian<float>  K;          // redefinition to avoid name clashing with OpenCV
+typedef CGAL::Simple_cartesian<double>  K;          // redefinition to avoid name clashing with OpenCV
 typedef K::Vector_3                     KVector;
 typedef K::Line_3                       KLine;
 typedef K::Plane_3                      KPlane;
